refactor : 쿼리 최적화 (N+1 문제) 해결
수정한 내용 :
- 명시적으로 모든 연관 관계를 지연 로딩으로 설정(디폴트는 toMany는 지연 로딩, toOne은 즉시 로딩)
1. getAuctionById 메서드 쿼리 최적화
- findAuctionById V2 메서드 추가
 -> Auction 조회 시 User를 fetch join하여 의도치 않은 쿼리 방지
 -> 쿼리 3개 => 쿼리 1개로 최적화

2. getProductByAucitonId 메서드 쿼리 최적화
- Product <-> ProductImage 간 양방향으로 매핑 관계 변경
 -> 기존은 Product -> ProductImage로의 단방향 매핑으로 조인 테이블 방식이었지만, 양방향 매핑 시 외래 키 매핑 방식으로 조인 테이블을 경유하지 않아 성능이 향상 됨.
- findByAuctionIdV2 메서드 추가
-> Product 조회 시 User와 컬렉션인 ProductImage까지 fetch join해서 가져 옴
-> 컬렉션은 지연 로딩 초기화 하는 방식이랑 트레이드 오프 고려 해야 함.

3. application.yml에 default batch size 설정
-> 엔티티 조회 시 지연 로딩을 in 쿼리로 묶어서 한 번에 로딩 하기 위함(쿼리 최적화)

추가한 메서드는 적용 안 했음
