+++ src/main/resources/application.yml
@@ -23,7 +23,7 @@ spring:
     properties:
       hibernate:
         format_sql: true
-#        default_batch_fetch_size: 100
+        default_batch_fetch_size: 100
 
     database-platform: org.hibernate.dialect.H2Dialect
   thymeleaf:
